Cleaned up the image scanning code. Added a version.
Went back to simple one pixel per tile scanning.  Works good enough for
images and since we can now read the meta data from the image, scanning
saved LiFE is trivial.

Added a version to the meta-data in case we want to change the meta
data format it in the future.