Added a method to send Mail
iOS7 on the iPad seems to have a serious flaw with sending messages
with an attachment.  The method fails and leaves your app with a black
view that cannot be dismissed.  So you have to quit the app and restart
it.  So I added a method to send mail and if we are on an iPad we send
an image via mail in lieu of sending a message.